+++ src/pages/NotasPage.tsx
@@ -4,43 +4,85 @@ import { NotaService } from "../services/NotasApi";
 import { toast } from "react-toastify";
 import Formulario from "../components/Formulario";
 import NotaForm from "../components/NotaForm";
+import Swal from "sweetalert2";
 
 const NotasPage = () => {
     const [datos, setDatos] = useState<Nota[]>([]);
+    const [datosOriginales, setDatosOriginales] = useState<Nota[]>([]);
     const [open, setOpen] = useState<boolean>(false);
     const [notaEditar, setNotaEditar] = useState<Nota | null>(null);
+    const [busqueda, setBusqueda] = useState<string>('');
 
     useEffect(() => {
-        NotaService.getNota().then((data) => setDatos(data));
+        actualizarDatos()
+        NotaService.getNota().then((data) => setDatosOriginales(data));
     }, [open]);
 
+    const actualizarDatos = () => {
+        NotaService.getNota().then((data) => setDatos(data));
+    };
+
+    const handleSearch = (event: React.ChangeEvent<HTMLInputElement>) => {
+        const consulta = event.target.value;
+        console.log(consulta);
+        setBusqueda(consulta);
+      
+        // Filtra los datos según la consulta de búsqueda
+        const datosFiltrados = datosOriginales.filter((dato) =>
+          dato.estudiante.toLowerCase().includes(consulta.toLowerCase())
+        );
+      
+        // Actualiza el estado con los datos filtrados o todos los datos originales
+        setDatos(consulta ? datosFiltrados : datosOriginales);
+      };
+
     const handleDelete = (idNota: string, idEstudiante: string) => {
-        NotaService.deleteNota(idNota, idEstudiante)
-            .then(response => {
-                console.log(response);
-                toast.success('La nota ha sido eliminado con éxito!');
-                NotaService.getNota().then((data) => setDatos(data));
-            })
-            .catch(error => {
-                console.error(error);
-                toast.error('Hubo un error al eliminar la nota');
-            });
+        Swal.fire({
+            title: 'Eliminar',
+            text: '¿Está seguro que desea borrar esta nota?',
+            icon: 'question',
+            showCancelButton: true,
+            confirmButtonText: 'Borrar',
+        }).then((result) => {
+            if (result.isConfirmed) {
+                NotaService.deleteNota(idNota, idEstudiante)
+                    .then(response => {
+                        console.log(response);
+                        toast.success('La nota ha sido eliminado con éxito!');
+                    })
+                    .catch(error => {
+                        console.error(error);
+                        toast.error('Hubo un error al eliminar la nota');
+                    });
+                actualizarDatos()
+                Swal.fire('Borrado', 'La nota ha sido eliminada con éxito', 'success');
+            } else {
+                Swal.fire('Cancelado', 'La nota no ha sido eliminada', 'info');
+            }
+        });
+
     };
 
     return (
         <>
-            <div className="p-10 flex justify-center w-full">
+            <div className="px-8 py-2 flex justify-end w-full">
+                <input
+                    type="text"
+                    placeholder="Buscar alumnos"
+                    value={busqueda}
+                    onChange={handleSearch}
+                />
                 <button
                     className="border border-neutral-300 rounded-lg py-1.5 px-10 my-2 bg-blue-500 hover:bg-blue-600 text-white "
-                    onClick={() => {setOpen(true); setNotaEditar(null)}}
+                    onClick={() => { setOpen(true); setNotaEditar(null) }}
                 >
                     Crear Nota
                 </button>
                 <Formulario open={open} onClose={() => setOpen(false)}>
-                    <NotaForm open={open} nota={notaEditar} onClose={() => setOpen(false)} />
+                    <NotaForm open={open} nota={notaEditar} onClose={() => setOpen(false)} actualizarDatos={actualizarDatos} />
                 </Formulario>
             </div>
-            <div className="flex flex-col p-5">
+            <div className="flex flex-col p-5 mx-5">
                 <div className="overflow-x-auto">
                     <div className="p-1.5 w-full inline-block align-middle">
                         <div className="overflow-x-auto border rounded-lg">
@@ -101,7 +143,7 @@ const NotasPage = () => {
                                                 {dato.porcentaje}
                                             </td>
                                             <td className="px-6 py-4 text-sm font-medium text-center">
-                                                <button className="text-green-500 hover:text-green-700" onClick={()=>{setNotaEditar(dato); setOpen(true);}}>
+                                                <button className="text-green-500 hover:text-green-700" onClick={() => { setNotaEditar(dato); setOpen(true); }}>
                                                     Editar
                                                 </button>
                                             </td>
@@ -116,6 +158,32 @@ const NotasPage = () => {
                             </table>
                         </div>
                     </div>
+                </div>
+            </div>
+            <div className="container mx-auto p-4">
+                <div className="grid grid-cols-3 gap-4">
+                    {datos.map((dato: any, index: number) => (
+                        <div key={index} className="bg-sky-200 m-5 shadow-lg rounded-lg border-8 border-blue-500">
+                            <div className="p-4 text-xl font-bold text-justify">Nombre: {dato.estudiante}</div>
+                            <div className="p-4 text-xl font-bold text-justify">Curso: {dato.curso}</div>
+                            <div className="p-4 text-xl font-bold text-justify">Nota: {dato.nota}</div>
+                            <div className="p-4 text-xl font-bold text-justify">Porcentaje: {dato.porcentaje}</div>
+                            <div className="flex justify-end p-2">
+                                <button
+                                    className="bg-green-500 hover:bg-green-600 text-white rounded-full px-4 h-8 mr-2"
+                                    onClick={() => { setNotaEditar(dato); setOpen(true); }}
+                                >
+                                    Editar
+                                </button>
+                                <button
+                                    className="bg-red-500 hover:bg-red-600 text-white rounded-full px-4 h-8"
+                                    onClick={() => handleDelete(dato.id, dato.estudiante_id)}
+                                >
+                                    Eliminar
+                                </button>
+                            </div>
+                        </div>
+                    ))}
                 </div>
             </div>
         </>
